Rewrote installation code
