+++ src/extension.ts
@@ -3,8 +3,7 @@ import { ExtensionConstants, LanguageServerConstants } from './constants';
 import { DafnyCommands } from './commands';
 import { DafnyLanguageClient } from './language/dafnyLanguageClient';
 import checkAndInformAboutInstallation from './startupCheck';
-import { DafnyInstaller, getLanguageServerExecutable, isConfiguredToInstallLatestDafny } from './language/dafnyInstallation';
-import { Messages } from './ui/messages';
+import { DafnyInstaller } from './language/dafnyInstallation';
 import createAndRegisterDafnyIntegration from './ui/dafnyIntegration';
 import { timeout } from './tools/timeout';
 import { fileIssueURL } from './ui/statusBarActionView';
@@ -45,24 +44,24 @@ class ExtensionRuntime {
   }
 
   public async initialize(): Promise<void> {
-    if(!this.installer.isCustomInstallation() && !await this.installer.isLanguageServerRuntimeAccessible()) {
-      if(!await this.installer.install()) {
-        window.showErrorMessage(Messages.Installation.Error);
-        return;
-      }
-    }
+    // if(!this.installer.isCustomInstallation() && !await this.installer.isLanguageServerRuntimeAccessible()) {
+    //   if(!await this.installer.install()) {
+    //     window.showErrorMessage(Messages.Installation.Error);
+    //     return;
+    //   }
+    // }
     await this.initializeClient();
-    if(!await this.updateDafnyIfNecessary(this.languageServerVersion!)) {
-      this.statusOutput.appendLine('Dafny initialization failed');
-      return;
-    }
-    await createAndRegisterDafnyIntegration(this.context, this.client!, this.languageServerVersion!);
+    // if(!await this.updateDafnyIfNecessary(this.languageServerVersion!)) {
+    //   this.statusOutput.appendLine('Dafny initialization failed');
+    //   return;
+    // }
+    await createAndRegisterDafnyIntegration(this.installer, this.client!, this.languageServerVersion!);
     commands.registerCommand(DafnyCommands.RestartServer, restartServer);
     this.statusOutput.appendLine('Dafny is ready');
   }
 
   private async initializeClient(): Promise<void> {
-    const exec = await getLanguageServerExecutable(this.context, [], []);
+    const exec = await this.installer.getLanguageServerExecutable([], []);
     this.statusOutput.appendLine(`Starting Dafny from ${exec.args![0]}`);
     await this.startClientAndWaitForVersion();
   }
@@ -80,31 +79,31 @@ class ExtensionRuntime {
     return version;
   }
 
-  private async updateDafnyIfNecessary(installedVersion: string): Promise<boolean> {
-    if(this.installer.isLatestKnownLanguageServerOrNewer(installedVersion)) {
-      return true;
-    }
-    if(this.installer.isCustomInstallation() || !isConfiguredToInstallLatestDafny()) {
-      window.showInformationMessage(
-        `${Messages.Installation.Outdated} ${installedVersion} < ${LanguageServerConstants.LatestVersion}`
-      );
-      return true;
-    }
-    await this.client!.stop();
-    if(!await this.installer.install()) {
-      window.showErrorMessage(Messages.Installation.Error);
-      return false;
-    }
-    await this.initializeClient();
-    return true;
-  }
+  // private async updateDafnyIfNecessary(installedVersion: string): Promise<boolean> {
+  //   if(this.installer.isLatestKnownLanguageServerOrNewer(installedVersion)) {
+  //     return true;
+  //   }
+  //   if(this.installer.isCustomInstallation() || !isConfiguredToInstallLatestDafny()) {
+  //     window.showInformationMessage(
+  //       `${Messages.Installation.Outdated} ${installedVersion} < ${LanguageServerConstants.LatestVersion}`
+  //     );
+  //     return true;
+  //   }
+  //   await this.client!.stop();
+  //   if(!await this.installer.install()) {
+  //     window.showErrorMessage(Messages.Installation.Error);
+  //     return false;
+  //   }
+  //   await this.initializeClient();
+  //   return true;
+  // }
 
   public async dispose(): Promise<void> {
     await this.client?.stop();
   }
 
   public async startClientAndWaitForVersion() {
-    this.client = this.client ?? await DafnyLanguageClient.create(this.context, this.statusOutput);
+    this.client = this.client ?? await DafnyLanguageClient.create(this.installer);
     this.client.start();
     await this.client.onReady();
     this.languageServerVersion = await this.getLanguageServerVersionAfterStartup();
@@ -123,7 +122,7 @@ class ExtensionRuntime {
     }
     this.context.subscriptions.splice(1);
     await this.initializeClient();
-    await createAndRegisterDafnyIntegration(this.context, this.client!, this.languageServerVersion!);
+    await createAndRegisterDafnyIntegration(this.installer, this.client!, this.languageServerVersion!);
     const issueURL = await fileIssueURL(this.languageServerVersion ?? '???', this.context);
     this.statusOutput.appendLine(
       'Dafny is ready again.\nIf you have time, please let us know why you needed to restart by filing an issue:\n'
