+++ src/ui/compileCommands.ts
@@ -7,7 +7,7 @@ import { DafnyCommands, VSCodeCommands } from '../commands';
 import Configuration from '../configuration';
 import { ConfigurationConstants } from '../constants';
 import { getDotnetExecutablePath } from '../dotnet';
-import { getCompilerRuntimePath } from '../language/dafnyInstallation';
+import { DafnyInstaller } from '../language/dafnyInstallation';
 import { Messages } from './messages';
 
 const CompileArg = '/compile';
@@ -15,17 +15,17 @@ const CompileAndRunArg = `${CompileArg}:3`;
 const OutputPathArg = '/out';
 
 export default class CompileCommands {
-  public static createAndRegister(context: ExtensionContext): CompileCommands {
-    context.subscriptions.push(
-      commands.registerCommand(DafnyCommands.Compile, () => compileAndRun(context, false, false)),
-      commands.registerCommand(DafnyCommands.CompileCustomArgs, () => compileAndRun(context, true, false)),
-      commands.registerCommand(DafnyCommands.CompileAndRun, () => compileAndRun(context, false, true))
+  public static createAndRegister(installer: DafnyInstaller): CompileCommands {
+    installer.context.subscriptions.push(
+      commands.registerCommand(DafnyCommands.Compile, () => compileAndRun(installer, false, false)),
+      commands.registerCommand(DafnyCommands.CompileCustomArgs, () => compileAndRun(installer, true, false)),
+      commands.registerCommand(DafnyCommands.CompileAndRun, () => compileAndRun(installer, false, true))
     );
     return new CompileCommands();
   }
 }
 
-async function compileAndRun(context: ExtensionContext, useCustomArgs: boolean, run: boolean): Promise<boolean> {
+async function compileAndRun(installer: DafnyInstaller, useCustomArgs: boolean, run: boolean): Promise<boolean> {
   const document = window.activeTextEditor?.document;
   if(document == null) {
     return false;
@@ -37,7 +37,7 @@ async function compileAndRun(context: ExtensionContext, useCustomArgs: boolean, 
   if(!await document.save()) {
     return false;
   }
-  const compilerCommand = await new CommandFactory(context, document.fileName, useCustomArgs, run).createCompilerCommand();
+  const compilerCommand = await new CommandFactory(installer, document.fileName, useCustomArgs, run).createCompilerCommand();
   if(compilerCommand == null) {
     return false;
   }
@@ -53,7 +53,7 @@ function runCommandInTerminal(command: string): void {
 
 class CommandFactory {
   public constructor(
-    private readonly context: ExtensionContext,
+    private readonly installer: DafnyInstaller,
     private readonly fileName: string,
     private readonly useCustomArgs: boolean,
     private readonly run: boolean
@@ -62,8 +62,8 @@ class CommandFactory {
   public async createCompilerCommand(): Promise<string | undefined> {
     const commandPrefix = this.getCommandPrefix();
     const { path: dotnetPath } = await getDotnetExecutablePath();
-    const compilerPath = await getCompilerRuntimePath(this.context);
     const compilerArgs = await this.getCompilerArgs();
+    const compilerPath = await this.installer.getCliExecutable(compilerArgs, compilerArgs);
     if(compilerArgs == null) {
       return undefined;
     }
@@ -75,22 +75,22 @@ class CommandFactory {
       ?? (os.type() === 'Windows_NT' ? '& ' : '');
   }
 
-  private async getCompilerArgs(): Promise<string | undefined> {
-    const configuredArgs = this.getConfiguredArgs().join(' ');
+  private async getCompilerArgs(): Promise<string[]> {
+    const configuredArgs = this.getConfiguredArgs();
     return this.useCustomArgs
-      ? await this.getCustomCompilerArgs(configuredArgs)
+      ? (await this.getCustomCompilerArgs(configuredArgs) ?? [])
       : configuredArgs;
   }
 
-  private async getCustomCompilerArgs(originalArgs: string): Promise<string | undefined> {
+  private async getCustomCompilerArgs(originalArgs: string[]): Promise<string[]> {
     const customArgs = await window.showInputBox({
-      value: originalArgs,
+      value: originalArgs.join(' '),
       prompt: Messages.Compiler.CustomArgumentsPrompt
     });
     if(customArgs == null) {
       window.showErrorMessage(Messages.Compiler.NoArgumentsGiven);
     }
-    return customArgs;
+    return customArgs?.split(' ') ?? [];
   }
 
   private getConfiguredArgs(): string[] {
