+++ src/ui/dafnyIntegration.ts
@@ -1,5 +1,3 @@
-import { ExtensionContext } from 'vscode';
-
 import { DafnyLanguageClient } from '../language/dafnyLanguageClient';
 import CompilationStatusView from './compilationStatusView';
 import CompileCommands from './compileCommands';
@@ -11,19 +9,20 @@ import RelatedErrorView from './relatedErrorView';
 import VerificationSymbolStatusView from './verificationSymbolStatusView';
 import Configuration from '../configuration';
 import { ConfigurationConstants } from '../constants';
+import { DafnyInstaller } from '../language/dafnyInstallation';
 
 export default async function createAndRegisterDafnyIntegration(
-  context: ExtensionContext,
+  installer: DafnyInstaller,
   languageClient: DafnyLanguageClient,
   languageServerVersion: string
 ): Promise<void> {
-  CounterexamplesView.createAndRegister(context, languageClient);
-  GhostDiagnosticsView.createAndRegister(context, languageClient);
-  const compilationStatusView = CompilationStatusView.createAndRegister(context, languageClient, languageServerVersion);
+  CounterexamplesView.createAndRegister(installer.context, languageClient);
+  GhostDiagnosticsView.createAndRegister(installer.context, languageClient);
+  const compilationStatusView = CompilationStatusView.createAndRegister(installer.context, languageClient, languageServerVersion);
   let symbolStatusView: VerificationSymbolStatusView | undefined = undefined;
   const serverSupportsSymbolStatusView = versionToNumeric('3.8.0') <= versionToNumeric(languageServerVersion);
   if(serverSupportsSymbolStatusView && Configuration.get<boolean>(ConfigurationConstants.LanguageServer.DisplayVerificationAsTests)) {
-    symbolStatusView = VerificationSymbolStatusView.createAndRegister(context, languageClient, compilationStatusView);
+    symbolStatusView = VerificationSymbolStatusView.createAndRegister(installer.context, languageClient, compilationStatusView);
   } else {
     if(serverSupportsSymbolStatusView) {
       compilationStatusView.registerAfter38Messages();
@@ -31,13 +30,16 @@ export default async function createAndRegisterDafnyIntegration(
       compilationStatusView.registerBefore38Messages();
     }
   }
-  VerificationGutterStatusView.createAndRegister(context, languageClient, symbolStatusView);
-  CompileCommands.createAndRegister(context);
-  RelatedErrorView.createAndRegister(context, languageClient);
-  await DafnyVersionView.createAndRegister(context, languageServerVersion);
+  VerificationGutterStatusView.createAndRegister(installer.context, languageClient, symbolStatusView);
+  CompileCommands.createAndRegister(installer);
+  RelatedErrorView.createAndRegister(installer.context, languageClient);
+  await DafnyVersionView.createAndRegister(installer, languageServerVersion);
 }
 
 export function versionToNumeric(version: string): number {
+  if(version.includes('nightly')) {
+    return Number.MAX_SAFE_INTEGER;
+  }
   const numbers = version.split('.').map(x => Number.parseInt(x));
-  return ((numbers[0] * 1000) + numbers[1]) * 1000 + numbers[2];
+  return ((numbers[0] * 1000) + (numbers[1] ?? 0)) * 1000 + (numbers[2] ?? 0);
 }
