+++ src/ui/dafnyVersionView.ts
@@ -4,26 +4,23 @@ import { promisify } from 'util';
 import { ExtensionContext, StatusBarAlignment, window } from 'vscode';
 
 import { LanguageServerConstants } from '../constants';
-import { getDotnetExecutablePath } from '../dotnet';
-import { getCompilerRuntimePath } from '../language/dafnyInstallation';
 import { enableOnlyForDafnyDocuments } from '../tools/visibility';
 import { version as ExtensionVersion } from '../../package.json';
+import { DafnyInstaller } from '../language/dafnyInstallation';
 
 const UnknownVersion = LanguageServerConstants.UnknownVersion;
-const CompilerVersionArg = '/version';
 const StatusBarPriority = 10;
 const execFileAsync = promisify(execFile);
 
-async function getTooltipText(context: ExtensionContext, languageServerVersion: string): Promise<string> {
-  const compilerVersion = await getCompilerVersion(context);
+async function getTooltipText(installer: DafnyInstaller, languageServerVersion: string): Promise<string> {
+  const compilerVersion = await getCompilerVersion(installer);
   return `Compiler: ${compilerVersion}\nLanguage Server: ${languageServerVersion}\nExtension: ${ExtensionVersion}`;
 }
 
-async function getCompilerVersion(context: ExtensionContext): Promise<string> {
-  const { path: dotnetPath } = await getDotnetExecutablePath();
-  const compilerPath = await getCompilerRuntimePath(context);
+async function getCompilerVersion(installer: DafnyInstaller): Promise<string> {
+  const compilerPath = await installer.getCliExecutable([ '--version' ], [ '/version' ]);
   try {
-    const { stdout } = await execFileAsync(dotnetPath, [ compilerPath, CompilerVersionArg ]);
+    const { stdout } = await execFileAsync(compilerPath.command, compilerPath.args);
     const version = /\d+\.\d+\.\d+\.\d+/.exec(stdout);
     return (version == null || version.length === 0) ? UnknownVersion : version[0];
   } catch(error: unknown) {
@@ -35,11 +32,11 @@ async function getCompilerVersion(context: ExtensionContext): Promise<string> {
 export default class DafnyVersionView {
   private constructor() {}
 
-  public static async createAndRegister(context: ExtensionContext, languageServerVersion: string): Promise<void> {
+  public static async createAndRegister(installer: DafnyInstaller, languageServerVersion: string): Promise<void> {
     const statusBarItem = window.createStatusBarItem(StatusBarAlignment.Right, StatusBarPriority);
     statusBarItem.text = languageServerVersion;
-    statusBarItem.tooltip = await getTooltipText(context, languageServerVersion);
-    context.subscriptions.push(
+    statusBarItem.tooltip = await getTooltipText(installer, languageServerVersion);
+    installer.context.subscriptions.push(
       enableOnlyForDafnyDocuments(statusBarItem),
       statusBarItem
     );
